Refactor: 인터페이스 주입
